requirements.txt
fastapi==0.115.6
httpx==0.26.0
Jinja2==3.1.5
joblib==1.3.2
multiprocess==0.70.15
numpy==1.26.4
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.18
regex==2023.10.3
requests==2.32.0
tqdm==4.66.3
uvicorn==0.23.2
virtualenv==20.26.6 
aiofiles==24.1.0
starlette>=0.40.0
# Web
beautifulsoup4==4.12.3
# PDF Parser
fastapi==0.115.6
uvicorn>=0.23.2
python-multipart==0.0.18
# PyMuPDF==1.24.0
PyPDF2>=3.0.1
pandas==2.2.3
# For Azure
azure-ai-documentintelligence==1.0.0b2
azure-common==1.1.28
azure-core==1.30.2
azure-ai-formrecognizer==3.3.3
#For Azure SQL
pymssql==2.3.2
# jaydebeapi==1.2.3 
# jpype1==1.5.2
# SQLAlchemy==2.0.38
# pyodbc==5.2.0
